requirements.txt
python-telegram-bot
google-genai
moviepy
requests